gui: file agenda view compiles and runs w/o issues
